Access control is done

+++ app/Http/Controllers/PostsController.php
@@ -12,6 +12,15 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -91,6 +100,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return redirect('/posts')->with('error','Unothorized access');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -108,6 +121,10 @@ class PostsController extends Controller
             'body'=>'required'
         ]);
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return redirect('/posts')->with('error','Unothorized access');
+        }
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -123,6 +140,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //check for correct user
+        if(auth()->user()->id!==$post->user_id){
+            return redirect('/posts')->with('error','Unothorized access');
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Deleted');
     }
